+++ ft_strncmp.c
@@ -6,14 +6,14 @@
 /*   By: kcisse <kcisse@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/20 15:08:31 by kcisse            #+#    #+#             */
-/*   Updated: 2024/05/21 19:06:48 by kcisse           ###   ########.fr       */
+/*   Updated: 2024/05/22 12:14:09 by kcisse           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
-int	ft_strncmp(const char *s1, const char *s2, size_t n)
+int	ft_strncmp(const char *s1, const char *s2, unsigned int n)
 {
-	size_t	i;
+	unsigned int	i;
 
 	i = 0;
 	while ((s1[i] || s2[i]) && i < n)
